# [b200] support trt-llm allreduce fuse rms_norm_add kernel