<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

I use flashinfer `27060628c32e1217e27564adf24e33273f4c8287` commit to test.

In b200:

Acc:

<img width="661" alt="图片" src="https://github.com/user-attachments/assets/5abed6d9-7013-47c3-be96-958d0caad823" />


main:

![图片](https://github.com/user-attachments/assets/5187f98f-6996-4b84-9f4b-5b26bfdbd454)

pr:

![图片](https://github.com/user-attachments/assets/e271d71b-54c1-4f1b-9c72-ca6074e26139)

13.5us->8.7us.

## end2end benchmark

```shell
python3 -m sglang.launch_server --model-path /dev/shm/DeepSeek-R1-FP4 --trust-remote-code --quantization modelopt_fp4 --tp 8 --enable-flashinfer-moe --speculative-algorithm=EAGLE --enable-flashinfer-allreduce-fusion
```
Refer to [trt-llm](https://github.com/NVIDIA/TensorRT-LLM/blob/16fc99391fdbe37e0d4f10ec2cc0fdc4cbb1cc75/tensorrt_llm/plugin/plugin.py#L708-L716) , I change `max_token_num` to 1024 and get a better result:

```markdown
  sglang python3 test/srt/parse_results.py dsv3_main.jsonl

Saved summary to: dsv3_main_summary.csv

+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |            133.166 |             133.166 |        176.653 |          153.561 |       226.394 |          7.337 |            7.599 |         8.060 |               133.166 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |             4.000 |            390.269 |             390.269 |        254.178 |          207.857 |       646.446 |          9.880 |            9.689 |        11.796 |                97.567 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |            16.000 |           1018.889 |            1018.889 |        275.613 |          179.920 |       643.978 |         14.551 |           14.247 |        20.015 |                63.681 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
➜  sglang python3 test/srt/parse_results.py dsv3_pr.jsonl  

Saved summary to: dsv3_pr_summary.csv

+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |            137.713 |             137.713 |        167.537 |          153.134 |       219.778 |          7.098 |            6.992 |         7.747 |               137.713 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |             4.000 |            409.847 |             409.847 |        208.332 |          156.417 |       537.090 |          9.388 |            9.090 |        11.284 |               102.462 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |            16.000 |           1055.758 |            1055.758 |        269.626 |          164.852 |       643.074 |         13.910 |           13.589 |        18.392 |                65.985 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
```

bs=1, 3.4%+
bs=8, 5%+
bs=16, 3.6+%


## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.