@BBuf 2025-06-28
b200

```markdown
➜  sglang python3 test/srt/parse_results.py dsv3_main.jsonl

Saved summary to: dsv3_main_summary.csv

+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |            128.105 |             128.105 |        193.663 |          195.547 |       246.901 |          7.617 |            7.431 |         8.464 |               128.105 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |             4.000 |            406.393 |             406.393 |        206.789 |          165.849 |       549.484 |          9.524 |            9.505 |        11.221 |               101.598 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |            16.000 |            998.480 |             998.480 |        271.431 |          224.228 |       692.032 |         14.818 |           14.609 |        21.982 |                62.405 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  3 |            32.000 |           1502.585 |            1502.585 |        331.783 |          223.811 |      1200.719 |         19.678 |           19.641 |        27.890 |                46.956 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
➜  sglang python3 test/srt/parse_results.py dsv3_pr.jsonl  

Saved summary to: dsv3_pr_summary.csv

+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|    |   max_concurrency |   input_throughput |   output_throughput |   mean_ttft_ms |   median_ttft_ms |   p99_ttft_ms |   mean_tpot_ms |   median_tpot_ms |   p99_tpot_ms |   per_user_throughput |
+====+===================+====================+=====================+================+==================+===============+================+==================+===============+=======================+
|  0 |             1.000 |            136.236 |             136.236 |        175.526 |          160.429 |       226.433 |          7.169 |            7.098 |         7.826 |               136.236 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  1 |             4.000 |            399.729 |             399.729 |        209.358 |          159.115 |       557.633 |          9.484 |            9.383 |        11.533 |                99.932 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  2 |            16.000 |            988.797 |             988.797 |        306.005 |          197.786 |       945.048 |         14.950 |           14.551 |        21.112 |                61.800 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+
|  3 |            32.000 |           1554.368 |            1554.368 |        368.550 |          178.469 |      1650.787 |         18.940 |           18.727 |        25.609 |                48.574 |
+----+-------------------+--------------------+---------------------+----------------+------------------+---------------+----------------+------------------+---------------+-----------------------+